add logger and move updateUser
