+++ common/user/index.ts
@@ -4,9 +4,6 @@
 import { pupil as Pupil, student as Student, screener as Screener } from '@prisma/client';
 import { prisma } from '../prisma';
 import { Prisma as PrismaTypes } from '@prisma/client';
-import { validateEmail } from '../../graphql/validators';
-import { updateZoomUser } from '../zoom/user';
-import { isZoomFeatureActive } from '../zoom/util';
 
 type Person = { id: number; isPupil?: boolean; isStudent?: boolean };
 
@@ -26,7 +23,6 @@ export type User = {
     studentId?: number;
     screenerId?: number;
 };
-
 export const userSelection = { id: true, firstname: true, lastname: true, email: true, active: true, lastLogin: true };
 
 export function getUserTypeAndIdForUserId(userId: string): [type: UserTypes, id: number] {
@@ -165,33 +161,6 @@ export async function queryUser<Select extends UserSelect>(user: User, select: S
     }
 
     throw new Error(`Unknown User(${user.userID})`);
-}
-
-export async function updateUser(userId: string, { email }: Partial<Pick<User, 'email'>>) {
-    const validatedEmail = validateEmail(email);
-    const user = await getUser(userId, /* active */ true);
-    if (user.studentId) {
-        if (isZoomFeatureActive()) {
-            await updateZoomUser(user);
-        }
-
-        return userForStudent(
-            (await prisma.student.update({
-                where: { id: user.studentId },
-                data: { email: validatedEmail },
-                select: userSelection,
-            })) as Student
-        );
-    }
-    if (user.pupilId) {
-        return userForPupil(
-            (await prisma.pupil.update({
-                where: { id: user.pupilId },
-                data: { email: validatedEmail },
-                select: userSelection,
-            })) as Pupil
-        );
-    }
 }
 
 export async function getUsers(userIds: User['userID'][]): Promise<User[]> {
